+++ packages/xchain-thorchain/src/types/client-types.ts
@@ -16,6 +16,9 @@ export type ExplorerUrls = {
 
 export type ExplorerUrl = Record<Network, string>
 
+export type ChainId = string
+export type ChainIds = Record<Network, ChainId>
+
 export type ThorchainClientParams = {
   clientUrl?: ClientUrl
   explorerUrls?: ExplorerUrls